# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateSupplier:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  models.Supplier:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateSupplier:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/supplier:
    get:
      consumes:
      - application/json
      description: Get a list of suppliers with optional filtering.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of suppliers
          schema:
            items:
              $ref: '#/definitions/models.Supplier'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a list of suppliers
      tags:
      - supplier
    post:
      consumes:
      - application/json
      description: Create a new supplier in the market system.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Supplier information
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/models.CreateSupplier'
      produces:
      - application/json
      responses:
        "201":
          description: Created supplier
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new supplier
      tags:
      - supplier
  /v1/supplier/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing supplier.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a supplier
      tags:
      - supplier
    get:
      consumes:
      - application/json
      description: Get supplier details by its ID.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Supplier details
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a supplier by ID
      tags:
      - supplier
    put:
      consumes:
      - application/json
      description: Update an existing supplier.
      parameters:
      - description: Authentication token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User password
        in: header
        name: Password
        required: true
        type: string
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated supplier information
        in: body
        name: supplier
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSupplier'
      produces:
      - application/json
      responses:
        "202":
          description: Updated supplier
          schema:
            $ref: '#/definitions/models.Supplier'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Supplier not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a supplier
      tags:
      - supplier
...
swagger: "2.0"